class name error ,fix it
